Add werror compile option to C++ (haog/gqe-python!147)
In gqe, we have set the option `add_compile_options("$<$<COMPILE_LANGUAGE:CXX>:-Werror>")`.

But we have not done this in gqe-python  . This is causing the CI to fail builds in gqe, that would normally be ignore in gqe-python.

This MR adds the option to gqe-python.

+++ gqe/benchmark/hardcoded/src/tpch/q22/task.cu
@@ -181,7 +181,7 @@ struct mark_join_functor {
                   std::is_same_v<identifier_type, int64_t>) {
       auto constexpr load_factor = 0.5;
       mark_join_map_type<identifier_type> mark_join_map{
-        customer_table_size,
+        static_cast<std::size_t>(customer_table_size),
         load_factor,
         empty_key_sentinel<identifier_type>,
         empty_value_sentinel,
